refactor(config-ui): move the error-handler to error as default
