+++ config-ui/src/error/components/index.ts
@@ -16,3 +16,4 @@
  */
 
 export * from './offline'
+export * from './default'
